// src/gql-client/__generated__/graphql.ts
export enum IssueState {
  /** An issue that has been closed */
  Closed = 'CLOSED',
  /** An issue that is still open */
  Open = 'OPEN'
}
// ...
export type GetRepositoryIssuesQuery = { __typename?: 'Query', repository?: { __typename?: 'Repository', issues: { __typename?: 'IssueConnection', totalCount: number, nodes?: Array<{ __typename?: 'Issue', state: IssueState, title: string, id: string, labels?: { __typename?: 'LabelConnection', nodes?: Array<{ __typename?: 'Label', color: string, name: string } | null> | null } | null, comments: { __typename?: 'IssueCommentConnection', totalCount: number } } | null> | null } } | null };

export type SearchIssuesByTitleOrBodyQueryVariables = Exact<{
  query: Scalars['String']['input'];
  numResults: Scalars['Int']['input'];
}>;


export type SearchIssuesByTitleOrBodyQuery = { __typename?: 'Query', search: { __typename?: 'SearchResultItemConnection', edges?: Array<{ __typename?: 'SearchResultItemEdge', node?: { __typename?: 'App' } | { __typename?: 'Discussion' } | { __typename?: 'Issue', state: IssueState, title: string, id: string, labels?: { __typename?: 'LabelConnection', nodes?: Array<{ __typename?: 'Label', color: string, name: string } | null> | null } | null, comments: { __typename?: 'IssueCommentConnection', totalCount: number }, author?: { __typename?: 'Bot', login: string } | { __typename?: 'EnterpriseUserAccount', login: string } | { __typename?: 'Mannequin', login: string } | { __typename?: 'Organization', login: string } | { __typename?: 'User', login: string } | null } | { __typename?: 'MarketplaceListing' } | { __typename?: 'Organization' } | { __typename?: 'PullRequest' } | { __typename?: 'Repository' } | { __typename?: 'User' } | null } | null> | null } };